nscd: disable caching for real
We noticed that nscd does cache things and sometimes causes weird bugs. Turning the TTL do 0 does not help in (some? all?) cases as the test was able to demonstrate even on the original commit where the TTL was lowered to 0.

This change ships a patch to completely short circuit nscd's cache functionality and a test that shows that `getaddrinfo` called
quickly twice causes proper resolver queries.

Could be used as a temporary fix for #135888 and #55276